requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-telegram-bot==20.7
mercadopago==2.3.0
apscheduler==3.10.4
